# reggie_OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "description": "瑞吉外卖接口文档",
    "version": "1.0",
    "title": "瑞吉外卖"
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/employee": {
      "post": {
        "tags": [
          "employee-controller"
        ],
        "summary": "save",
        "operationId": "saveUsingPOST_3",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "employee",
            "description": "employee",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Employee",
              "originalRef": "Employee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»",
              "originalRef": "R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "employee-controller"
        ],
        "summary": "update",
        "operationId": "updateUsingPUT_2",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "employee",
            "description": "employee",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Employee",
              "originalRef": "Employee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»",
              "originalRef": "R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/employee/login": {
      "post": {
        "tags": [
          "employee-controller"
        ],
        "summary": "login",
        "operationId": "loginUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "employee",
            "description": "employee",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Employee",
              "originalRef": "Employee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«Employee»",
              "originalRef": "R«Employee»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/employee/logout": {
      "post": {
        "tags": [
          "employee-controller"
        ],
        "summary": "logout",
        "operationId": "logoutUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«string»",
              "originalRef": "R«string»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/employee/page": {
      "get": {
        "tags": [
          "employee-controller"
        ],
        "summary": "page",
        "operationId": "pageUsingGET_2",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "name",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«Page»",
              "originalRef": "R«Page»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/employee/{id}": {
      "get": {
        "tags": [
          "employee-controller"
        ],
        "summary": "getById",
        "operationId": "getByIdUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R«Employee»",
              "originalRef": "R«Employee»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Employee": {
      "type": "object",
      "properties": {
        "createTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "createUser": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "idNumber": {
          "type": "string",
          "refType": null
        },
        "name": {
          "type": "string",
          "refType": null
        },
        "password": {
          "type": "string",
          "refType": null
        },
        "phone": {
          "type": "string",
          "refType": null
        },
        "sex": {
          "type": "string",
          "refType": null
        },
        "status": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "updateUser": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "username": {
          "type": "string",
          "refType": null
        }
      },
      "title": "Employee"
    },
    "OrderItem": {
      "type": "object",
      "properties": {
        "asc": {
          "type": "boolean",
          "refType": null
        },
        "column": {
          "type": "string",
          "refType": null
        }
      },
      "title": "OrderItem"
    },
...
    "Page": {
      "type": "object",
      "properties": {
        "countId": {
          "type": "string",
          "refType": null
        },
        "current": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "hitCount": {
          "type": "boolean",
          "refType": null
        },
        "maxLimit": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "optimizeCountSql": {
          "type": "boolean",
          "refType": null
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem",
            "originalRef": "OrderItem"
          },
          "refType": "OrderItem"
        },
        "pages": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "records": {
          "type": "array",
          "items": {
            "type": "object"
          },
          "refType": "object"
        },
        "searchCount": {
          "type": "boolean",
          "refType": null
        },
        "size": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "refType": null
        }
      },
      "title": "Page"
    },
...
    "R«Employee»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/Employee",
          "originalRef": "Employee"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«Employee»"
    },
...
    "R«Page»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "$ref": "#/definitions/Page",
          "originalRef": "Page"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«Page»"
    },
...
    "R«string»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "string"
        },
        "map": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "R«string»"
    },
...
  }
}